Stickfemale_and_Stickmale_Porn
Erosion-Brosion Pornography goes here. Artworks completed by MWK/RWW, in Firealpaca.  (If you have any concerns on pornography on GitHub due its policies, interest in my understanding after I read the terms, that GitHub allows sexual content as long as it is not for mere surprise or shock, that is has purpose, not spammy, and other suches. These images have purposing in the Oterhes, as well as canonical value and other suches.)

Drawn by the author, artwork of Stickmale and Stickfemale fucking/Stickmale below and Stickfemale above giving a cute blowjob.